requirements.txt
fastapi==0.119.0
uvicorn[standard]==0.38.0
asyncpg==0.30.0
pytest==8.4.2
pytest-asyncio==1.2.0
httpx==0.28.1
python-multipart==0.0.20
PyJWT==2.8.0
passlib==1.7.4
bcrypt==4.0.1